+++ CHANGELOG.rst
@@ -3,6 +3,8 @@ Membranecurvature CHANGELOG
 
 ---
 
+* Enabled ``curvature_on`` kwarg to calculate curvature as ``per-frame`` average or as
+  ``'average_surface'``.
 * Added ``surface_method='binning_nearest'`` as a new surface method (PR #246)
 * Feature: Added periodic edge padding for the binning surface method (PR #238)
 * Refactored binning surface method to use ``np.add.at`` for improved performance (PR #240)